Fix createStringAutocomplete module resolution for production build (#32)
* Fix createStringAutocomplete module resolution for production build

- Split createAutocomplete and createStringAutocomplete into separate files
- Export both functions individually to fix production bundling issue
- Update Dockerfile cache bust for clean rebuild
- Resolves 'Cannot read properties of undefined' error in Fly.io deployment



* Use direct imports for createStringAutocomplete to fix production bundling

- Import createStringAutocomplete directly from its file instead of through barrel export
- This ensures proper module resolution in production builds
- Fixes undefined error when accessing the function

* Update export pattern for autocomplete functions

- Use named export at end of file instead of export function declaration
- Matches common pattern for better tree-shaking
- Should resolve production bundling issues

* fix: Convert autocomplete factory functions to React components

- Created ModelAutocomplete component to replace createAutocomplete factory
- Created StringAutocomplete component to replace createStringAutocomplete factory
- Updated all component usages to use new direct components with model prop
- Fixed production build issues with Next.js bundling of factory functions
- Tested build locally and confirmed it compiles successfully

This resolves the "Cannot read properties of undefined" errors in production
by using standard React components instead of higher-order factory functions.



* fix: Remove unused factory function files

- Deleted createAutocomplete.tsx and createStringAutocomplete.tsx
- These factory functions have been replaced with direct React components
- Removes potential bundling conflicts in production build



* Fix formatting issues for CI

* Fix TypeScript any type errors for CI

* Fix Prettier formatting for CI

* Formatting

---------
