+++ src/components/ui/index.ts
@@ -48,5 +48,6 @@ export * from "@/components/ui/BaseDataGrid"
 export * from "@/components/ui/filters"
 export * from "@/components/ui/lists/Manager"
 export * from "@/components/ui/SearchInput"
-export * from "@/components/ui/createAutocomplete"
+export * from "@/components/ui/ModelAutocomplete"
+export * from "@/components/ui/StringAutocomplete"
 export * from "@/components/ui/CampaignGuard"
